rename err_file_modal to election_results_reporting_file_modal
